Commented out the base_path variable
This base_path variable was defaulted to /mnt/volume in the inventory.sample file.  This has lead to some issues with novice users in which this path doesn't exist on their file system or the polkadot user didn't have access to it.  I believe the default path should be used in all cases and this be used by exception.